refactored the carousel component
